refactor(compressor): move compressor file to `model` package
Added `DispatcherProvider.kt` to get rid of hardcoded dispatchers.
